Cleanup
Fixed some bugs I found in testing.
The parsing logic I originally wrote was working, but hard to read. Implemented a cleaner solution of string.split() with the delimiters.
Fixed a unit test that was not setup correctly.
Added catch case when user enters no values in the calculator.
Tweaked how we identify when a customer delimiter has been added, and when to record those delimiters.